update to streaming market data endpoint
