+++ customers.go
@@ -104,22 +104,3 @@ func (c *Client) GetMyAccount(accountNumber string) (Account, *http.Response, er
 
 	return customersRes.Account, resp, nil
 }
-
-// Returns the appropriate quote streamer endpoint, level and identification token.
-// for the current customer to receive market data.
-func (c *Client) GetQuoteStreamerTokens() (QuoteStreamerTokenAuthResult, *http.Response, error) {
-	path := "/quote-streamer-tokens"
-
-	type customerResponse struct {
-		Streamer QuoteStreamerTokenAuthResult `json:"data"`
-	}
-
-	customersRes := new(customerResponse)
-
-	resp, err := c.request(http.MethodGet, path, nil, nil, customersRes)
-	if err != nil {
-		return QuoteStreamerTokenAuthResult{}, resp, err
-	}
-
-	return customersRes.Streamer, resp, nil
-}
